Moved file storage related validation into media.py
Also made media.py more like a flask extension by adding an init_app function which registers the blueprint and validator.
